[experiment] fast list cleanup (#169)
* fast-cleanup

* generalize cleanup

+++ src/utils/component.ts
@@ -208,9 +208,10 @@ export function destroyElementSync(
     | Array<ComponentReturnType | Node>
     | null
     | null[],
+  skipDom = false
 ) {
   if (isArray(component)) {
-    component.forEach((component) => destroyElementSync(component));
+    component.forEach((component) => destroyElementSync(component, skipDom));
   } else {
     if (isEmpty(component)) {
       return;
@@ -219,6 +220,9 @@ export function destroyElementSync(
     if ($nodes in component) {
       if (component.ctx !== null) {
         runDestructorsSync(component.ctx);
+      }
+      if (skipDom) {
+        return;
       }
       try {
         destroyNodes(component[$nodes]);
@@ -229,6 +233,9 @@ export function destroyElementSync(
         );
       }
     } else {
+      if (skipDom) {
+        return;
+      }
       destroyNode(component);
     }
   }
@@ -261,9 +268,10 @@ export async function destroyElement(
     | Array<ComponentReturnType | Node>
     | null
     | null[],
+  skipDom = false,
 ) {
   if (isArray(component)) {
-    await Promise.all(component.map((component) => destroyElement(component)));
+    await Promise.all(component.map((component) => destroyElement(component, skipDom)));
   } else {
     if (component === null) {
       return;
@@ -274,6 +282,9 @@ export async function destroyElement(
         runDestructors(component.ctx, destructors);
         await Promise.all(destructors);
       }
+      if (skipDom) {
+        return;
+      }
       try {
         destroyNodes(component[$nodes]);
       } catch (e) {
@@ -283,6 +294,9 @@ export async function destroyElement(
         );
       }
     } else {
+      if (skipDom) {
+        return;
+      }
       await destroyNode(component);
     }
   }
